+++ tests/testthat/test-data_summary.R
@@ -282,7 +282,7 @@ test_that("data_summary, extra functions", {
 
 
 test_that("data_summary, bayestestR::ci", {
-  skip_if_not_installed("bayestesR")
+  skip_if_not_installed("bayestestR")
   data(mtcars)
   out <- data_summary(
     mtcars,
@@ -290,20 +290,270 @@ test_that("data_summary, bayestestR::ci", {
     ci = bayestestR::ci(mpg),
     by = c("am", "gear")
   )
+  expect_named(out, c("am", "gear", "mean_value", "CI", "CI_low", "CI_high"))
   expect_snapshot(out)
-  expect_error(
-    data_summary(
-      mtcars,
-      mw = mean(mpg),
-      test = bayestestR::ci(mpg),
-      yolo = c(mean(mpg), sd(mpg)),
-      by = c("am", "gear")
-    ),
-    regex = "Each expression"
+  out <- data_summary(
+    mtcars,
+    mw = mean(mpg),
+    test = bayestestR::ci(mpg),
+    yolo = c(mean(mpg), sd(mpg)),
+    by = c("am", "gear")
+  )
+  expect_named(
+    out,
+    c("am", "gear", "mw", "CI", "CI_low", "CI_high", "yolo_1", "yolo_2")
   )
 })
 
 test_that("no warning when variable name and function in global env clash, #583", {
   dat <- data.frame(rt = 1:10)
   expect_silent(data_summary(dat, rt = mean(rt)))
+})
+
+
+test_that("allow multiple columns for expressions", {
+  set.seed(123)
+  d <- data.frame(
+    x = rnorm(100, 1, 1),
+    y = rnorm(100, 2, 2),
+    groups = rep(1:4, each = 25)
+  )
+
+  out <- data_summary(
+    d,
+    quant_x = quantile(x, c(0.25, 0.75)),
+    quant_y = quantile(y, c(0.25, 0.75)),
+    suffix = c("Q1", "Q3")
+  )
+  expect_equal(
+    out$quant_xQ1,
+    0.50615,
+    tolerance = 1e-3,
+    ignore_attr = TRUE
+  )
+  expect_named(out, c("quant_xQ1", "quant_xQ3", "quant_yQ1", "quant_yQ3"))
+
+  # automatic suffixes
+  out <- data_summary(
+    d,
+    quant_x = quantile(x, c(0.25, 0.75)),
+    quant_y = quantile(y, c(0.1, 0.9)),
+    suffix = NULL
+  )
+  expect_named(out, c("quant_x25%", "quant_x75%", "quant_y10%", "quant_y90%"))
+
+  # use own suffix only for one expression - other expressions are
+  # suffixed with `_1`, `_2`, etc.
+  out <- data_summary(
+    d,
+    quant_x = quantile(x, c(0.25, 0.75)),
+    quant_y = quantile(y, c(0.25, 0.5, 0.75)),
+    mean_x = mean(x),
+    suffix = list(quant_y = c("_Q1", "_Q2", "_Q3")),
+    by = "groups"
+  )
+  expect_named(
+    out,
+    c(
+      "groups",
+      "quant_x25%",
+      "quant_x75%",
+      "quant_y_Q1",
+      "quant_y_Q2",
+      "quant_y_Q3",
+      "mean_x"
+    )
+  )
+
+  set.seed(123)
+  d <- data.frame(
+    x = rnorm(100, 1, 1),
+    y = rnorm(100, 2, 2),
+    w = rnorm(100, 3, 0.5),
+    z = rnorm(100, 4, 3),
+    groups = rep(1:4, each = 25)
+  )
+
+  out <- data_summary(
+    d,
+    quant_x = quantile(x, c(0.25, 0.75)),
+    mean_x = mean(x),
+    quant_y = quantile(y, c(0.25, 0.5, 0.75))
+  )
+  expect_equal(
+    out,
+    data.frame(
+      `quant_x25%` = 0.50615,
+      `quant_x75%` = 1.69182,
+      mean_x = 1.09041,
+      `quant_y25%` = 0.39779,
+      `quant_y50%` = 1.54834,
+      `quant_y75%` = 2.93569
+    ),
+    tolerance = 1e-3,
+    ignore_attr = TRUE
+  )
+
+  out <- data_summary(
+    d,
+    quant_x = quantile(x, c(0.25, 0.75)),
+    mean_x = mean(x),
+    fivenum = fivenum(y)
+  )
+  expect_equal(
+    out,
+    data.frame(
+      `quant_x25%` = 0.50615,
+      `quant_x75%` = 1.69182,
+      mean_x = 1.09041,
+      fivenum_1 = -2.10649,
+      fivenum_2 = 0.36539,
+      fivenum_3 = 1.54834,
+      fivenum_4 = 2.96837,
+      fivenum_5 = 8.48208
+    ),
+    tolerance = 1e-3,
+    ignore_attr = TRUE
+  )
+
+  out <- data_summary(
+    d,
+    quant_x = quantile(x, c(0.25, 0.75)),
+    mean_x = mean(x),
+    quant_y = quantile(y, c(0.25, 0.5, 0.75)),
+    suffix = list(quant_y = c("_Q1", "_Q2", "_Q3"))
+  )
+  expect_equal(
+    out,
+    data.frame(
+      `quant_x25%` = 0.50615,
+      `quant_x75%` = 1.69182,
+      mean_x = 1.09041,
+      quant_y_Q1 = 0.39779,
+      quant_y_Q2 = 1.54834,
+      quant_y_Q3 = 2.93569
+    ),
+    tolerance = 1e-3,
+    ignore_attr = TRUE
+  )
+
+  out <- data_summary(
+    d,
+    quant_x = quantile(x, c(0.25, 0.75)),
+    mean_x = mean(x),
+    quant_y = quantile(y, c(0.25, 0.5, 0.75)),
+    suffix = list(quant_x = c("Q1", "Q3"), quant_y = c("_Q1", "_Q2", "_Q3"))
+  )
+  expect_equal(
+    out,
+    data.frame(
+      quant_xQ1 = 0.50615,
+      quant_xQ3 = 1.69182,
+      mean_x = 1.09041,
+      quant_y_Q1 = 0.39779,
+      quant_y_Q2 = 1.54834,
+      quant_y_Q3 = 2.93569
+    ),
+    tolerance = 1e-3,
+    ignore_attr = TRUE
+  )
+
+  out <- data_summary(
+    d,
+    quant_x = quantile(x, c(0.25, 0.5)),
+    quant_w = quantile(w, c(0.25, 0.5)),
+    quant_y = quantile(y, c(0.25, 0.5)),
+    quant_z = quantile(z, c(0.25, 0.5)),
+    suffix = c("_Q1", "_Q2")
+  )
+  expect_equal(
+    out,
+    data.frame(
+      quant_x_Q1 = 0.50615,
+      quant_x_Q2 = 1.06176,
+      quant_w_Q1 = 2.73435,
+      quant_w_Q2 = 3.01796,
+      quant_y_Q1 = 0.39779,
+      quant_y_Q2 = 1.54834,
+      quant_z_Q1 = 1.81187,
+      quant_z_Q2 = 3.98947
+    ),
+    tolerance = 1e-3,
+    ignore_attr = TRUE
+  )
+
+  # errors ------------------------------------------------------------------
+
+  expect_error(
+    data_summary(
+      d,
+      quant_x = quantile(x, c(0.25, 0.75)),
+      mean_x = mean(x),
+      quant_y = quantile(y, c(0.25, 0.5, 0.75)),
+      suffix = list(quant_xy = c("_Q1", "_Q2", "_Q3"))
+    ),
+    regex = "Names of `suffix` must match the names",
+    fixed = TRUE
+  )
+
+  expect_error(
+    data_summary(
+      d,
+      quant_x = quantile(x, c(0.25, 0.75)),
+      mean_x = mean(x),
+      quant_y = quantile(y, c(0.25, 0.5, 0.75)),
+      suffix = list(c("Q1", "Q3"), "mean", c("_Q1", "_Q2", "_Q3"))
+    ),
+    regex = "All elements of `suffix` must have names.",
+    fixed = TRUE
+  )
+
+  expect_error(
+    data_summary(
+      d,
+      quant_x = quantile(x, c(0.25, 0.75)),
+      mean_x = mean(x),
+      quant_y = quantile(y, c(0.25, 0.5, 0.75)),
+      suffix = c("_Q1", "_Q2", "_Q3")
+    ),
+    regex = "Argument `suffix` must have the same length",
+    fixed = TRUE
+  )
+
+  expect_error(
+    data_summary(
+      d,
+      quant_x = quantile(x, c(0.25, 0.75)),
+      mean_x = mean(x),
+      quant_y = quantile(y, c(0.25, 0.5, 0.75)),
+      suffix = list(quant_x = c("_Q1", "_Q2", "_Q3"))
+    ),
+    regex = "Argument `suffix` must have the same length",
+    fixed = TRUE
+  )
+
+  expect_error(
+    data_summary(
+      d,
+      quant_x = quantile(x, c(0.25, 0.75)),
+      mean_x = mean(x),
+      quant_y = quantile(y, c(0.25, 0.5, 0.75)),
+      suffix = list(quant_x = c("Q1", "Q3"), quant_y = c("_Q1", "_Q2", "_Q2"))
+    ),
+    regex = "All suffixes for a single expression must be unique",
+    fixed = TRUE
+  )
+
+  expect_error(
+    data_summary(
+      d,
+      quant_x = quantile(x, c(0.25, 0.5)),
+      quant_w = quantile(w, c(0.25, 0.5)),
+      quant_y = quantile(y, c(0.25, 0.5)),
+      quant_z = quantile(z, c(0.25, 0.5)),
+      suffix = c("_Q1", "_Q2", "_Q3")
+    ),
+    regex = "Argument `suffix` must have the same length",
+    fixed = TRUE
+  )
 })